+++ opencore_legacy_patcher/efi_builder/firmware.py
@@ -216,6 +216,14 @@ class BuildFirmware:
         if not "CPU Generation" in smbios_data.smbios_dictionary[self.model]:
             return
 
+        # APFS check
+        # The macOS 26 APFS EFI driver's FileVault 2 implementation is broken, and
+        # must be replaced with the macOS 15 APFS EFI driver.
+        logging.info("- Enabling macOS 26 FileVault 2 support")
+        self.config["UEFI"]["APFS"]["EnableJumpstart"] = False
+        shutil.copy(self.constants.sequoia_apfs_driver_path, self.constants.drivers_path)
+        support.BuildSupport(self.model, self.constants, self.config).get_efi_binary_by_path("apfs_aligned.efi", "UEFI", "Drivers")["Enabled"] = True
+
         # Exfat check
         if smbios_data.smbios_dictionary[self.model]["CPU Generation"] < cpu_data.CPUGen.sandy_bridge.value:
             # Sandy Bridge and newer Macs natively support ExFat